requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
requests==2.31.0
lxml==4.9.3
pandas==2.1.3
Pillow==10.1.0
google-generativeai==0.3.1
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
dj-database-url==2.1.0
curl-cffi==0.7.0